Maybe: Add static const object none
That is an instance of type Maybe<void>, used to represent the absence
of a value.

Additionally, added support for operator== and operator!= for
Maybe<void>, hence none, with the semantics:

Maybe<void> x;
Maybe<void> y;

x == y => true
x != y => false

Given that it doesn't hold a value.

Therefore, none will be always equal to itself.